. single instance imageviewer . only pics from current session are shown
git-svn-id: https://svn.code.sf.net/p/qterm/code/trunk/qterm@266 bd9efd21-7b0e-0410-bb59-ddbc1b2ac886

+++ qterm/qtermcanvas.cpp
@@ -11,57 +11,34 @@
 #include <qwmatrix.h>
 #include <qpopupmenu.h>
 #include <qmessagebox.h>
+#include <qstring.h>
 
-extern QString getSaveFileName(const QString&, QWidget*);
 extern QString fileCfg;
+extern QString getSaveFileName(const QString&, QWidget*);
 
 QTermCanvas::QTermCanvas(QWidget *parent, const char *name, WFlags f)
   :QScrollView(parent, name, f)
 {
-
-	//dirty trick
-	if (f == 0)
-		bEmbed = true;
-	else
-		bEmbed = false;
-
-	m_pMenu = new QPopupMenu(this);
-	m_pMenu->insertItem( tr("zoom 1:1"), this, SLOT(oriSize()), Key_Z );
-	m_pMenu->insertItem( tr("fit window"), this, SLOT(fitWin()), Key_X );
-	m_pMenu->insertSeparator();	
-	m_pMenu->insertItem( tr("zoom in"), this, SLOT(zoomIn()), Key_Equal );
-	m_pMenu->insertItem( tr("zoom out"), this, SLOT(zoomOut()), Key_Minus );
-	if (!bEmbed)
-		m_pMenu->insertItem( tr("fullscreen"), this, SLOT(fullScreen()), Key_F );
-	m_pMenu->insertSeparator();	
-	m_pMenu->insertItem( tr("rotate CW 90"), this, SLOT(cwRotate()), Key_BracketRight );
-	m_pMenu->insertItem( tr("rotate CCW 90"), this, SLOT(ccwRotate()), Key_BracketLeft );
-
-	if (!bEmbed) {
-		m_pMenu->insertSeparator();
-		m_pMenu->insertItem( tr("save as"), this, SLOT(saveImage()), Key_S );
-		m_pMenu->insertItem( tr("copy to"), this, SLOT(copyImage()), Key_C );
-		m_pMenu->insertItem( tr("silent copy"), this, SLOT(silentCopy()), Key_S+SHIFT );
-		m_pMenu->insertItem( tr("delete"), this, SLOT(deleteImage()), Key_D );
-
-		m_pMenu->insertSeparator();
-		m_pMenu->insertItem( tr("exit"), this, SLOT(close()), Key_Q );
-	}
-
 	bFitWin=true;
 	
 	label = new QLabel(viewport());
-	label->setScaledContents(true);
+	label->setScaledContents(false);
 	label->setAlignment(AlignCenter);
 	label->setText("No Preview Available");
 	addChild(label);
-	resize(200,100);
+
+	setHScrollBarMode(AlwaysOff);
+	setVScrollBarMode(AlwaysOff);
 
 }
 QTermCanvas::~QTermCanvas()
 {
 	delete label;
-	delete m_pMenu;
+}
+
+void QTermCanvas::setFixed(bool fixed)
+{
+	bFixed = fixed;
 }
 
 void QTermCanvas::oriSize()
@@ -99,48 +76,51 @@ void QTermCanvas::ccwRotate()
 	rotateImage(-90);
 }
 
-void QTermCanvas::fullScreen()
-{
-	if(!isFullScreen())
-		showFullScreen();
-	else
-		showNormal();
-}
 
-void QTermCanvas::loadImage(QString name)
+void QTermCanvas::loadImage(const QString& name)
 {
+	label->setText("Loading...");
 	img.load(name);
 	if(!img.isNull())
 	{
 
 		strFileName = name;
-		setCaption(QFileInfo(name).fileName());
+       		
+		parentWidget()->setCaption(QFileInfo(name).fileName());
 
-		bFitWin=true;
-
-		QSize szView(img.size());
-		szView.scale(640,480,QSize::ScaleMin);
 		
-		if(szView.width()<img.width())
+		if( bFixed || parentWidget()->isFullScreen() )
 		{
-			szImage = szView;
-			label->setPixmap(scaleImage(szImage));
-			if (!bEmbed)
-				resize(szView*1.1);
+			label->setPixmap(QPixmap(img));
+			szImage = img.size();
+			fitWin();
 		}
 		else
 		{
-			szImage = img.size();
-			label->setPixmap(QPixmap(img));
-			if (!bEmbed)
-				resize(szImage+QSize(5,5));
+
+			QSize szView(img.size());
+			szView.scale(640,480,QSize::ScaleMin);
+			if(szView.width()<img.width()) // increase window size
+			{
+				szImage = szView;
+				label->setPixmap(scaleImage(szImage));
+				emit resizeWindow(szView);
+			}
+			else	// decrease window size
+			{
+				szImage = img.size();
+				label->setPixmap(QPixmap(img));
+				emit resizeWindow(szImage);
+			}
+
 		}
-		if (bEmbed)
-			fitWin();
-		
+
 	}
 	else
+	{
+		label->setText("Failed to display");
 		qWarning("cant load image "+name);
+	}
 }
 
 void QTermCanvas::resizeImage(float ratio)
@@ -155,14 +135,18 @@ void QTermCanvas::resizeImage(float ratio)
 		return;
 	szImage = szImg;
 
-	if(!isFullScreen() && !bEmbed)		
-		resize(szImage*1.1);
+	if(!parentWidget()->isFullScreen() && !bFixed)		
+		emit resizeWindow(szImage);
 	else
 		adjustSize(QSize(visibleWidth(), visibleHeight()));
 }
 
 void QTermCanvas::rotateImage(float ang)
 {
+
+	if(label->pixmap()==NULL)
+		return;
+
 	QWMatrix wm;
 	
 	wm.rotate(ang);
@@ -176,6 +160,9 @@ void QTermCanvas::rotateImage(float ang)
 
 void QTermCanvas::copyImage()
 {
+	if(label->pixmap()==NULL)
+		return;
+
 	QFileInfo fi(strFileName);
 	QString strSave = getSaveFileName(fi.fileName(), this);
 
@@ -198,6 +185,9 @@ void QTermCanvas::copyImage()
 
 void QTermCanvas::silentCopy()
 {
+	if(label->pixmap()==NULL)
+		return;
+
 	// save it to $savefiledialog
 	QTermConfig conf(fileCfg);
 	QString strPath = QString::fromLocal8Bit(
@@ -249,6 +239,9 @@ void QTermCanvas::moveImage(float dx, float dy)
 
 void QTermCanvas::saveImage()
 {
+	if(label->pixmap()==NULL)
+		return;
+
 	QFileInfo fi(strFileName);
 	QString strSave = getSaveFileName(fi.fileName(),this);
 
@@ -261,14 +254,11 @@ void QTermCanvas::saveImage()
 
 void QTermCanvas::deleteImage()
 {
-	QFile::remove(strFileName);
-	close();
-}
+	if(label->pixmap()==NULL)
+		return;
 
-void QTermCanvas::closeEvent(QCloseEvent *ce)
-{
-	if (!bEmbed)
-		delete this;
+	QFile::remove(strFileName);
+	emit filesChanged();
 }
 
 void QTermCanvas::viewportResizeEvent(QResizeEvent *re)
@@ -276,49 +266,10 @@ void QTermCanvas::viewportResizeEvent(QResizeEvent *re)
 	adjustSize(re->size());
 }
 
-void QTermCanvas::contentsMousePressEvent(QMouseEvent *me)
-{
-/* remove this to avoid click by mistake
-	if(me->button()&LeftButton)
-	{
-		close();
-		return;
-	}
-*/
-	if(me->button()&RightButton)
-	{
-		m_pMenu->popup(me->globalPos());
-	}
-}
-
-void QTermCanvas::keyPressEvent(QKeyEvent *ke)
-{
-	switch(ke->key())
-	{
-		case Qt::Key_Escape:
-			if(isFullScreen())
-				showNormal();
-			else
-				close();
-			break;
-		case Qt::Key_Left:
-			moveImage(-0.05,0);
-			break;
-		case Qt::Key_Right:
-			moveImage(0.05,0);
-			break;
-		case Qt::Key_Up:
-			moveImage(0,-0.05);
-			break;
-		case Qt::Key_Down:
-			moveImage(0,0.05);
-			break;
-
-	}
-}
 
 void QTermCanvas::adjustSize(const QSize& szView)
 {
+	// occupy whole visible area if nothing
 	if(label->pixmap()==NULL)
 	{
 		label->resize(visibleWidth(), visibleHeight());
@@ -326,6 +277,7 @@ void QTermCanvas::adjustSize(const QSize& szView)
 	}
 
 	QSize szImg=szImage;
+
 
 	if(bFitWin)
 	{
@@ -343,13 +295,28 @@ void QTermCanvas::adjustSize(const QSize& szView)
 
 	szImage = szImg;
 
+
+	// I decide to set on/off scrollbar manually because:
+	// when ScrollBarMode=Auto, I will receive one more
+	// viewportResizeEvent, since it tries to compute
+	if(szView.width()<szImg.width())
+		setHScrollBarMode(AlwaysOn);
+	else
+		setHScrollBarMode(AlwaysOff);
+	
+	if(szView.height()<szImg.height())
+		setVScrollBarMode(AlwaysOn);
+	else
+		setVScrollBarMode(AlwaysOff);
+
+	
 	if(bFitWin)
 		resizeContents(szView.width(), szView.height());
 	else
-		resizeContents(szImage.width()*1.1,szImage.height()*1.1);
+		resizeContents(szImage.width(),szImage.height());
+
 
 	moveChild(label, x>0?x/2:0, y>0?y/2:0); 
-	label->resize(szImage);
 
 	label->setPixmap(scaleImage(szImage));
